Did final refactoring of the components
