+++ packages/bot-web-ui/src/components/dashboard/bot-builder/toolbox/toolbox.tsx
@@ -1,61 +1,11 @@
 import React from 'react';
 import classNames from 'classnames';
-import { Field as FormField, Formik, Form, FieldProps } from 'formik';
-import { Input, Icon, Text } from '@deriv/components';
+import { Icon, Text } from '@deriv/components';
 import { localize } from '@deriv/translations';
 import { ToolboxItems } from './toolbox-items';
 import { connect } from '../../../../stores/connect';
 import RootStore from 'Stores/index.js';
-
-type TSearchBox = {
-    is_search_loading: boolean;
-    onSearch: () => void;
-    onSearchBlur: () => void;
-    onSearchClear: (
-        param: (field: string, value: number | string, shouldValidate?: boolean | undefined) => void
-    ) => void;
-    onSearchKeyUp: (param: () => void) => void;
-};
-
-type TFormValues = { [key: string]: string };
-
-const SearchBox = ({ is_search_loading, onSearch, onSearchBlur, onSearchClear, onSearchKeyUp }: TSearchBox) => (
-    <div className='db-toolbox__search'>
-        <Formik initialValues={{ search: '' }} onSubmit={onSearch}>
-            {({ submitForm, values: { search }, setFieldValue }) => (
-                <Form>
-                    <FormField name='search'>
-                        {({ field }: FieldProps<string, TFormValues>) => (
-                            <Input
-                                {...field}
-                                className='db-toolbox__search-field'
-                                type='text'
-                                name='search'
-                                placeholder={localize('Search')}
-                                onKeyUp={() => onSearchKeyUp(submitForm)}
-                                onFocus={submitForm}
-                                onBlur={onSearchBlur}
-                                leading_icon={
-                                    (search &&
-                                        (is_search_loading ? (
-                                            <div className='loader' />
-                                        ) : (
-                                            <Icon
-                                                icon='IcCloseCircle'
-                                                onClick={() => onSearchClear(setFieldValue)}
-                                                color='secondary'
-                                            />
-                                        ))) ||
-                                    (!search && <Icon icon='IcSearch' />)
-                                }
-                            />
-                        )}
-                    </FormField>
-                </Form>
-            )}
-        </Formik>
-    </div>
-);
+import SearchBox from './search-box';
 
 type TToolbox = {
     hasSubCategory: (param: HTMLCollection) => boolean;
@@ -108,12 +58,20 @@ const Toolbox = ({
                 <div className='db-toolbox__header'>
                     <div className='db-toolbox__title' onClick={() => setOpen(!is_open)}>
                         {localize('Blocks menu')}
-                        <span className={classNames('chevron', { active: is_open })}>
+                        <span
+                            className={classNames('db-toolbox__title-chevron', {
+                                'db-toolbox__title-chevron--active': is_open,
+                            })}
+                        >
                             <Icon icon='IcChevronDownBold' />
                         </span>
                     </div>
                 </div>
-                <div className={classNames('db-toolbox__content-wrapper', { active: is_open })}>
+                <div
+                    className={classNames('db-toolbox__content-wrapper', {
+                        'db-toolbox__content-wrapper--active': is_open,
+                    })}
+                >
                     <SearchBox
                         is_search_loading={is_search_loading}
                         onSearch={onSearch}
