+++ packages/bot-web-ui/src/components/dashboard/dashboard-components/index.scss
@@ -24,7 +24,8 @@
 
         .dc-tabs {
             &__active-line {
-                background: none;
+                background-color: var(--general-main-1);
+                transition: none;
             }
 
             &__active {
@@ -50,8 +51,51 @@
                     margin-right: 0.8rem;
                 }
             }
+            &__dashboard-cards-wrapper {
+                display: flex;
+            }
 
-            &__content_group {
+            &__dashboard-card {
+                margin-right: 4.2rem;
+                display: flex;
+                flex-direction: column;
+
+                svg {
+                    margin-bottom: 0.8rem;
+
+                    &:hover {
+                        transition: all 100ms ease-in;
+                        box-shadow: rgba(0, 0, 0, 0.35) 0px 5px 15px;
+                        border-radius: 5px;
+                        font-weight: 700;
+                    }
+                }
+
+                span {
+                    text-align: center;
+                }
+
+                &:hover {
+                    cursor: pointer;
+                }
+                &:last-child {
+                    margin-right: 0;
+                }
+                &-image {
+                    background: var(--general-section-6);
+                }
+                &-label {
+                    font-size: 14px;
+                    font-weight: 400;
+                    line-height: 20px;
+                }
+
+                img {
+                    margin-bottom: 0.8rem;
+                }
+            }
+
+            &__content-group {
                 display: flex;
                 flex-direction: column;
                 justify-content: center;
@@ -59,58 +103,18 @@
                 height: calc(100vh - 20rem);
                 margin-top: 1rem;
 
-                &_heading {
+                .heading {
                     font-size: 20px;
                     font-weight: 700;
                     line-height: 30px;
                     margin-bottom: 16px;
                 }
 
-                &_description {
+                .description {
                     font-size: 16px;
                     font-weight: 400;
                     line-height: 24px;
                     margin-bottom: 5rem;
-                }
-
-                &_tiles {
-                    display: flex;
-
-                    &_block {
-                        margin-right: 4.2rem;
-                        display: flex;
-                        flex-direction: column;
-                        cursor: pointer;
-                        &:last-child {
-                            margin-right: 0;
-                        }
-                        svg {
-                            margin-bottom: 0.8rem;
-                            &:hover {
-                                transition: all 100ms ease-in;
-                                box-shadow: rgba(0, 0, 0, 0.35) 0px 5px 15px;
-                                border-radius: 5px;
-                                font-weight: 700;
-                            }
-                        }
-
-                        span {
-                            text-align: center;
-                        }
-                    }
-
-                    img {
-                        margin-bottom: 0.8rem;
-                    }
-
-                    &_content {
-                        font-size: 14px;
-                        font-weight: 400;
-                        line-height: 20px;
-                    }
-                    &_images {
-                        background: var(--general-section-6);
-                    }
                 }
             }
         }
